// Aplikacija/Angular app/PartyMaker/src/app/models/returnUser.model.ts
export class ReturnUser{
  password: string;
  email: string;
  licnoIme: string;
  prezime: string;
  admin: boolean;

  // constructor(password:string, email:string, licnoIme:string, prezime:string, admin:boolean){
  //   this.password=password;
  //   this.email=email;
  //   this.licnoIme=licnoIme;
  //   this.prezime=prezime;
  //   this.admin=admin;
  // }
}